move header hash out of DefaultHeaderMap
